Fix some CI build and test issues
This change udpates some of the out of date github workflow action
helpers to the latest versions to get past deprecation errors.

As well, the '--theme Default' option is passed into the Application
during tests in order to bypass the theme selector dialog which most
tests are not expecting to be present and thus hang waiting for the main
dialog to be active.

Finally, the branches_panel test is updated to use the insecure
git//github.com/stinb/gitahead-test.git url instead of the default
git@github.com:stinb/gitahead-test.git url to avoid existing issues with
ssh, specifically when run in a container without ssh keys present (such
as in the github workflow actions environment).